+++ src/components/editor/editor-pane/tool-bar/editor-preferences/editor-preference-property.ts
@@ -1,0 +1,13 @@
+/*
+SPDX-FileCopyrightText: 2020 The HedgeDoc developers (see AUTHORS file)
+
+SPDX-License-Identifier: AGPL-3.0-only
+*/
+
+export enum EditorPreferenceProperty {
+  KEYMAP = 'keyMap',
+  THEME = 'theme',
+  INDENT_WITH_TABS = 'indentWithTabs',
+  INDENT_UNIT = 'indentUnit',
+  SPELL_CHECK = 'spellcheck'
+}