+++ src/redux/editor/types.ts
@@ -11,12 +11,14 @@ import { EditorMode } from '../../components/editor/app-bar/editor-view-mode'
 export enum EditorConfigActionType {
   SET_EDITOR_VIEW_MODE = 'editor/mode/set',
   SET_SYNC_SCROLL = 'editor/syncScroll/set',
-  SET_EDITOR_PREFERENCES = 'editor/preferences/set'
+  MERGE_EDITOR_PREFERENCES = 'editor/preferences/merge',
+  SET_LIGATURES = 'editor/preferences/setLigatures'
 }
 
 export interface EditorConfig {
   editorMode: EditorMode;
   syncScroll: boolean;
+  ligatures: boolean
   preferences: EditorConfiguration
 }
 
@@ -26,6 +28,10 @@ export interface EditorConfigActions extends Action<EditorConfigActionType> {
 
 export interface SetEditorSyncScrollAction extends EditorConfigActions {
   syncScroll: boolean
+}
+
+export interface SetEditorLigaturesAction extends EditorConfigActions {
+  ligatures: boolean
 }
 
 export interface SetEditorConfigAction extends EditorConfigActions {